Fix the event timestamp size (32 bits vs 64 bits).

+++ src/common/net/Deserializer.h
@@ -21,7 +21,7 @@ namespace net {
  */
 struct BaseEventData {
   quint32 id;
-  quint64 timestamp;
+  quint32 timestamp;
 };
 
 /**
